requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
google-generativeai>=0.8.0
groq>=0.9.0
rpy2>=3.5.0
